rename HeroSection
